[TAN-1155] Introduce new report_data_grouping feature
